backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==14.1
deepgram-sdk==3.8.3
groq==0.11.0
python-dotenv==1.0.1
python-multipart==0.0.9
pypdf2==3.0.1
pillow==10.4.0
pytesseract==0.3.13
langchain==0.3.7
langchain-community==0.3.5
langchain-text-splitters==0.3.2
langchain-huggingface==0.1.2
faiss-cpu==1.9.0
sentence-transformers==3.3.1
requests==2.32.3